requirements.txt
robotframework>=3.0.4
junitparser==2.0
PyYAML>=3.13

### Dev
mock>=2.0.0
invoke>=1.1.1
coverage>=5.1
testfixtures>=6.14.1 # needed for large dict comparisons to make sense of them
pytest>=7.4.2
pytest-cov>=4.1.0
docutils>=0.16   # needed to generate library documentation with libdoc
Pygments>=2.6.1  # this one too
twine>=3.1.1 # needed for releasing to pypi
build>=0.6.0 # needed for building the distribution
